+++ tests/integration/setup.ts
@@ -5,6 +5,8 @@ import { resetFactoryCounters } from "../factories";
 export const prisma = new PrismaClient();
 
 const TABLES = [
+  "Opportunity",
+  "AgentProfile",
   "WorkflowRun",
   "QaLog",
   "CallTranscript",